Move unit tests for Time class to a separate test_time.py file
